test: osd/TestOSDMap.cc: fix Clang complain about promotion

+++ src/test/osd/TestOSDMap.cc
@@ -441,7 +441,7 @@ TEST(PGTempMap, basic)
   pg_t a(1,1);
   for (auto i=3; i<1000; ++i) {
     pg_t x(i, 1);
-    m.set(x, {i});
+    m.set(x, {static_cast<int>(i)});
   }
   pg_t b(2,1);
   m.set(a, {1, 2});
